+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- `--porcelain` flag that outputs script friendly output
 
 ## [0.1.2] - 2016-05-30
 ### Added